HotelWise/CloudFunctions/DB_FIX_GCP/requirements.txt
functions-framework==3.*
google-cloud-storage
pandas
fsspec
gcsfs
openpyxl
pyarrow
